Service 로직 리팩터링 완료

+++ backend/src/main/java/fullcare/backend/project/service/ProjectService.java
@@ -1,14 +1,18 @@
 package fullcare.backend.project.service;
 
 import fullcare.backend.global.State;
+import fullcare.backend.global.errorcode.MemberErrorCode;
 import fullcare.backend.global.errorcode.ProjectErrorCode;
 import fullcare.backend.global.exceptionhandling.exception.CompletedProjectException;
 import fullcare.backend.global.exceptionhandling.exception.EntityNotFoundException;
-import fullcare.backend.global.exceptionhandling.exception.InvalidAccessException;
+import fullcare.backend.global.exceptionhandling.exception.InvalidDateRangeException;
+import fullcare.backend.global.exceptionhandling.exception.UnauthorizedAccessException;
 import fullcare.backend.member.domain.Member;
+import fullcare.backend.member.repository.MemberRepository;
 import fullcare.backend.project.domain.Project;
 import fullcare.backend.project.dto.request.ProjectCreateRequest;
 import fullcare.backend.project.dto.request.ProjectUpdateRequest;
+import fullcare.backend.project.dto.response.ProjectDetailResponse;
 import fullcare.backend.project.dto.response.ProjectListResponse;
 import fullcare.backend.project.dto.response.ProjectSimpleListResponse;
 import fullcare.backend.project.repository.ProjectRepository;
@@ -35,10 +39,16 @@ import java.util.stream.Collectors;
 public class ProjectService {
 
     private final ProjectRepository projectRepository;
+    private final MemberRepository memberRepository;
+
     private final S3Service s3Service;
 
+
+    // ! 꼭 Project 클래스타입으로 반환해야하는가?(테스트 데이터 생성을 위해 클래스 타입으로 반환중)
     @Transactional
-    public Project createProject(Member member, ProjectCreateRequest request) {
+    public Project createProject(Long memberId, ProjectCreateRequest request) {
+        Member findMember = memberRepository.findById(memberId).orElseThrow(() -> new EntityNotFoundException(MemberErrorCode.MEMBER_NOT_FOUND));
+
         Project newProject = Project.createNewProject()
                 .title(request.getTitle())
                 .description(request.getDescription())
@@ -48,34 +58,87 @@ public class ProjectService {
                 .imageUrl(request.getImageUrl())
                 .build();
 
-        newProject.addMember(member, new ProjectMemberType(ProjectMemberRoleType.리더, ProjectMemberPositionType.미정));
+        newProject.addMember(findMember, new ProjectMemberType(ProjectMemberRoleType.리더, ProjectMemberPositionType.미정));
         return projectRepository.save(newProject);
     }
 
     @Transactional
-    public void updateProject(Long projectId, ProjectUpdateRequest request) {
-        Project findProject = findSimpleProject(projectId);
-        s3Service.delete(findProject.getImageUrl());
+    public void updateProject(Long projectId, Long memberId, ProjectUpdateRequest request) {
 
-        findProject.updateAll(request.getTitle(), request.getDescription(), request.getState(),
-                request.getStartDate(), request.getEndDate(), request.getImageUrl());
+        try {
+            ProjectMember findProjectMember = isProjectAvailable(projectId, memberId, false);
+
+            // ! 프로젝트에 소속되어있지만 리더가 아니라서 수정 권한이 없음 -> 403 Forbidden
+            if (!(findProjectMember.isLeader())) {
+                throw new UnauthorizedAccessException(ProjectErrorCode.UNAUTHORIZED_MODIFY);
+            }
+
+            if (!request.getStartDate().isBefore(request.getEndDate())) {
+                throw new InvalidDateRangeException(ProjectErrorCode.INVALID_DATE_RANGE);
+            }
+
+            Project findProject = findProjectMember.getProject();
+            s3Service.delete(findProject.getImageUrl());
+            findProject.updateAll(request.getTitle(), request.getDescription(), request.getState(),
+                    request.getStartDate(), request.getEndDate(), request.getImageUrl());
+
+        } catch (CompletedProjectException completedProjectException) {
+            throw new CompletedProjectException(ProjectErrorCode.INVALID_MODIFY);
+        }
     }
 
     @Transactional
-    public void deleteProject(Long projectId) {
-        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
+    public void deleteProject(Long projectId, Long memberId) {
 
-        if (project.getImageUrl() != null) {
-            s3Service.delete(project.getImageUrl());
+        try {
+            ProjectMember findProjectMember = isProjectAvailable(projectId, memberId, false);
+
+            // ! 프로젝트에 소속되어있지만 리더가 아니라서 삭제 권한이 없음 -> 403 Forbidden
+            if (!(findProjectMember.isLeader())) {
+                throw new UnauthorizedAccessException(ProjectErrorCode.UNAUTHORIZED_DELETE);
+            }
+
+            Project findProject = findProjectMember.getProject();
+            if (findProject.getImageUrl() != null) {
+                s3Service.delete(findProject.getImageUrl());
+            }
+
+            projectRepository.delete(findProject);
+
+        } catch (CompletedProjectException completedProjectException) {
+            throw new CompletedProjectException(ProjectErrorCode.INVALID_DELETE);
+        }
+    }
+
+    @Transactional
+    public void completeProject(Long projectId, Long memberId) {
+
+        ProjectMember findProjectMember = isProjectAvailable(projectId, memberId, false);
+
+        // ! 프로젝트에 소속되어있지만 리더가 아니라서 상태 수정 권한이 없음 -> 403 Forbidden
+        if (!findProjectMember.isLeader()) {
+            throw new UnauthorizedAccessException(ProjectErrorCode.UNAUTHORIZED_COMPLETE);
         }
 
-        projectRepository.deleteById(projectId);
+        Project findProject = findProjectMember.getProject();
+        findProject.complete();
     }
 
-    @Transactional
-    public void completeProject(Long projectId) {
-        Project findProject = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
-        findProject.complete();
+    public ProjectDetailResponse findProjectDetail(Long projectId, Long memberId) {
+
+        ProjectMember findProjectMember = isProjectAvailable(projectId, memberId, true);
+        Project findProject = findProjectMember.getProject();
+
+        ProjectDetailResponse projectDetailResponse = ProjectDetailResponse.builder()
+                .title(findProject.getTitle())
+                .description(findProject.getDescription())
+                .startDate(findProject.getStartDate())
+                .endDate(findProject.getEndDate())
+                .imageUrl(findProject.getImageUrl())
+                .state(findProject.getState())
+                .build();
+
+        return projectDetailResponse;
     }
 
     public Project findProject(Long projectId) {
@@ -108,7 +171,7 @@ public class ProjectService {
     }
 
     // ! todo readonly 옵션으로 데이터 읽기 용도로만 호출할 경우 프로젝트가 완료되어도 검증 통과
-    public ProjectMember isProjectAvailable(Long projectId, Long memberId, boolean readOnly) {
+    public ProjectMember isProjectAvailable(Long projectId, Long memberId, boolean readOnlyAfterProjectComplete) {
         // * 프로젝트가 존재하는 검증
         Project findProject = projectRepository.findProjectWithPMAndMemberById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
 
@@ -117,10 +180,10 @@ public class ProjectService {
         // ! 사용자가 프로젝트에 소속되어있지 않다면, 예외를 던진다.
         ProjectMember findProjectMember = findProject.getProjectMembers().stream()
                 .filter(pmm -> pmm.getMember().getId().equals(memberId)).findAny()
-                .orElseThrow(() -> new InvalidAccessException(ProjectErrorCode.INVALID_ACCESS));
+                .orElseThrow(() -> new UnauthorizedAccessException(ProjectErrorCode.UNAUTHORIZED_ACCESS));
 
         // * 프로젝트가 완료되었는지 검증
-        if (findProject.isCompleted() && !readOnly) {
+        if (findProject.isCompleted() && !readOnlyAfterProjectComplete) {
             throw new CompletedProjectException(ProjectErrorCode.PROJECT_COMPLETED);
         }
